+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added PPS metrics on the Encoded Transform when used.
+
 ### Removed
 
 ### Changed
 
 - Refactored DefaultUserAgentParser to utilize userAgentData when possible, e.g. to get Chrome minor version. Refactored all previous use of UAParser to use DefaultUserAgentParser. This does not change existing dimensions on event reporting. 
 - Updated to Typescript 5.x and node >= 20
+- Refactored Encoded Transform management into it's own component for better support of non-redundant audio transforms.
 
 ### Fixed